--- configs/cylinder3d/cylinder3d_4xb4_3x_semantickitti.py
@@ -1,37 +1,0 @@
-_base_ = [
-    '../_base_/datasets/semantickitti.py', '../_base_/models/cylinder3d.py',
-    '../_base_/default_runtime.py'
-]
-
-# optimizer
-# This schedule is mainly used by models on nuScenes dataset
-lr = 0.001
-optim_wrapper = dict(
-    type='OptimWrapper',
-    optimizer=dict(type='AdamW', lr=lr, weight_decay=0.01))
-
-train_cfg = dict(type='EpochBasedTrainLoop', max_epochs=36, val_interval=1)
-val_cfg = dict(type='ValLoop')
-test_cfg = dict(type='TestLoop')
-
-# learning rate
-param_scheduler = [
-    dict(
-        type='LinearLR', start_factor=0.001, by_epoch=False, begin=0,
-        end=1000),
-    dict(
-        type='MultiStepLR',
-        begin=0,
-        end=36,
-        by_epoch=True,
-        milestones=[30],
-        gamma=0.1)
-]
-
-# Default setting for scaling LR automatically
-#   - `enable` means enable scaling LR automatically
-#       or not by default.
-#   - `base_batch_size` = (8 GPUs) x (4 samples per GPU).
-# auto_scale_lr = dict(enable=False, base_batch_size=32)
-
-default_hooks = dict(checkpoint=dict(type='CheckpointHook', interval=5))
